Wallet

+++ FawryAppProject/src/javaApp/Payment.java
@@ -1,8 +1,7 @@
 package javaApp;
 
-public interface Payment {
-	
-	public void pay(int amount);
+public interface Payment {	
+	public void pay(double amount);
 	public static final Discount disc = new Discount();
 	
 }
